Fix connection issue in MySQLi class
Updated the MySQL connection logic within src/pool/classes/Database/Driver/MySQLi.php to better handle the absence of scheme in hostname and correctly use 'pass' for password. This resolves the connection problems faced when 'user' or 'pass' was not present in hostname. It also falls back to $username and $password if 'user' and 'pass' are not available in $connectionParameters.