Add Curl utility class to project
A new utility class, Curl, was added to the 'POOL' namespace contained in the pool/utils directory. This class provides two static methods - isValidUrl and downloadFile. These methods are used to validate a given URL and download a file from a given URL respectively. This implementation provides a centralized way to perform these commonly used operations and increases reusability and consistency across the project.