Add getGUIClassFolder method to GUI_Module
A new method getGUIClassFolder was added to GUI_Module to enhance path handling for template, css and javascript files. It uses the Str::sliceAfter utility from the newly imported Str utility class, improving code readability and ease of maintainability. This change tackles the issue of inefficient and unclear code for dealing with paths in the class.